+++ forui/lib/widgets.dart
@@ -1,10 +1,10 @@
 /// The Forui widgets and their corresponding styles.
 library forui.widgets;
 
-export 'src/widgets/badge/badge.dart' hide FBadgeContent, Variant;
-export 'src/widgets/button/button.dart' hide FButtonContent, Variant;
-export 'src/widgets/card/card.dart' hide FCardContent;
-export 'src/widgets/dialog/dialog.dart' hide FDialogContent, FHorizontalDialogContent, FVerticalDialogContent;
+export 'src/widgets/badge/badge.dart' hide Variant;
+export 'src/widgets/button/button.dart' hide Variant;
+export 'src/widgets/card/card.dart';
+export 'src/widgets/dialog/dialog.dart';
 export 'src/widgets/header/header.dart';
 export 'src/widgets/tabs/tabs.dart';
 export 'src/widgets/text_field/text_field.dart';