+++ forui/test/src/widgets/separator_golden_test.dart
@@ -50,7 +50,7 @@ void main() {
 
           await expectLater(
             find.byType(TestScaffold),
-            matchesGoldenFile('separator/$name-$orientation.png'),
+            matchesGoldenFile('separator/$name-$orientation-separator.png'),
           );
         });
       }